6-size.c

+++ 0x00-hello_world/6-size.c
@@ -15,9 +15,9 @@ int main(void)
 	float f;
 
 	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of an int: %lu byte(s)\n", (unsigned int)sizeof(b));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
 	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
 	printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of float: %lu byte(s)\n", (unsigned int)sizeof(f));
+	printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
